Bug fix between pawns and kings
With the checking system it checked all available squares that pieces could move(of opposite color to king) and kings(of opposite color to pieces) could not move into those squares. The exception with this was pawns because pawns can move forward but cannot take other pieces the same way. they can only take diagonally if a piece of opposite color is there.