Add computer again
This time what i did is that the computer goes through their pieces and filters out the ones that can move. then randomly chooses one of those pieces and randomly chooses one of the possible moves. it then gets the position of the piece and the move and uses the click handler to make its move.